requirements.txt
aiofiles
annotated-types
click
colorama
confluent-kafka
contourpy
cycler
dict-toolbox
fonttools
ghp-import
importlib_metadata
importlib_resources
Jinja2
kiwisolver
lazy-object-proxy
Markdown
MarkupSafe
matplotlib
mergedeep
mkdocs
mkdocs-get-deps
msgpack
nest-asyncio
numpy
packaging
pathspec
pillow
pip-review
platformdirs
pop
pop-config
pop-loop
pyasn1
pydantic
pydantic_core
pyparsing
PySimpleGUI
python-dateutil
PyYAML
pyyaml_env_tag
rend
rsa
six
sniffio
sumolib
toml
traci
typing_extensions
watchdog
zipp
pandas
requests
paramiko
#pip install git+https://github.com/dpkp/kafka-python.git
kafka-python~=2.0.3.dev0